+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 1.2.0
+
+* R/W Strings / Byte Arrays
+
 ## 1.1.0
 
-* Packed Varints~~~~
+* Packed Varints
 
 ## 1.0.0